scrapers/boldtv/requirements.txt
attrs==21.4.0
Automat==20.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==37.0.2
cssselect==1.1.0
filelock==3.7.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.6.0
itemloaders==1.0.4
jmespath==1.0.0
lxml==4.8.0
mysql-connector==2.2.9
mysql-connector-python==8.0.29
parsel==1.6.0
Protego==0.2.1
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pyOpenSSL==22.0.0
queuelib==1.6.2
requests==2.27.1
requests-file==1.5.1
Scrapy==2.6.1
service-identity==21.1.0
six==1.16.0
tldextract==3.3.0
Twisted==22.4.0
typing-extensions==4.2.0
urllib3==1.26.9
w3lib==1.22.0
zope.interface==5.4.0
